% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dailyrep.R
\name{dailyrep}
\alias{dailyrep}
\title{Download and Report Daily GreenFeed Data}
\arguments{
\item{User}{User name to log in to GreenFeed}

\item{Pass}{Password to log in to GreenFeed}

\item{Exp}{Study name}

\item{Unit}{The unit number(s) of the GreenFeed. If multiple units, they should be separated by a comma}

\item{Start_Date}{Start date of the study}

\item{End_Date}{End date of the study. If not specified, the current date will be used}

\item{Dir}{Directory to save the output file. If not specified, the current working directory will be used}

\item{Plot_opt}{Type of gas to plot: All, or CH4, CO2, O2, H2. If not specified, only CH4 will be processed}

\item{RFID_file}{The file that contains the RFID of the animals in the study}
}
\value{
A .csv file with daily data from GreenFeed unit(s) and
    a PDF report with a description of daily data
}
\description{
`dailyrep()` downloads daily data using an application programming interface (API)
    and generates a PDF report to verify the correct functionality of GreenFeed unit(s).
    This function retrieves data based on the specified parameters, processes it,
    and outputs a summary report in PDF format. The report includes essential metrics and plots
    to ensure that the GreenFeed units are operating as expected.
}
\examples{
\dontshow{if (has_credentials()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
# Please replace "your_username" and "your_password" with your actual GreenFeed credentials.
User <- Sys.getenv("API_USER")
Pass <- Sys.getenv("API_PASS")
Exp <- "StudyName"
Unit <- 1
Start_Date <- "2023-01-01"
End_Date <- Sys.Date()
Dir <- getwd()
Plot_opt <- "All"

# Example without RFID_file (by default NA)
dailyrep(User, Pass, Exp, Unit, Start_Date, End_Date, Dir, Plot_opt)

\dontshow{\}) # examplesIf}
}
