% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{download_data}
\alias{download_data}
\title{Download data from 'API'}
\usage{
download_data(
  user,
  pass,
  d,
  type = 2,
  unit = NULL,
  start_date = NULL,
  end_date = NULL
)
}
\value{
A dataframe
}
\description{
Download all types of 'GreenFeed' from 'API'
}
\examples{
\dontshow{if (has_credentials()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
# Please replace "your_username" and "your_password" with your actual 'GreenFeed' credentials.
# By default, the function downloads the preliminary 'GreenFeed' data,
# if raw data is needed use options: "feed", "rfid", or "cmds"

download_data(
  user = "your_username",
  pass = "your_password",
  d = "visits",
  type = 2,
  unit = c(304, 305),
  start_date = "2024-01-01",
  end_date = Sys.Date()
)
\dontshow{\}) # examplesIf}
}
\keyword{internal}
