% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{filter_within_range}
\alias{filter_within_range}
\title{Detect outliers in data using mean and standard deviation}
\usage{
filter_within_range(v, cutoff)
}
\arguments{
\item{v}{A vector with data}

\item{cutoff}{A threshold or cutoff value that defines the range (e.g., 2.5)}
}
\value{
A logical vector of the same length as `v`, where each element is `TRUE` if the corresponding value in `v` falls within the specified range, and `FALSE` otherwise.
}
\description{
`filter_within_range()` is a helper function to detect
    outliers in data using the mean and sd
}
