% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/finalrep.R
\name{finalrep}
\alias{finalrep}
\title{Report Final GreenFeed Data}
\arguments{
\item{Exp}{Study name}

\item{Unit}{The unit number(s) of the GreenFeed. If multiple units, they should be separated by a comma.}

\item{Start_Date}{Start date of the study}

\item{End_Date}{End date of the study}

\item{Final_report}{List with final report/s generated by C-Lock}

\item{Plot_opt}{Type of gas to plot: All, or combination of them "CH4, CO2", or "CH4", and so on. If not specified, only CH4 will be process}

\item{RFID_file}{The file that contains the RFID of the animals in the study}
}
\value{
A PDF report with description of final GreenFeed data.
}
\description{
`finalrep()` generates a final report for GreenFeed data, summarizing key metrics
    and providing a comprehensive overview of the collected data.
    This report includes analysis and visualizations to assist in evaluating
    the performance and outcomes of the GreenFeed units.
}
\examples{
# Create a Final Report in PDF format from the finalized data received from C-Lock Inc.
# Define parameters:
Experiment <- "StudyName"

# Note that Unit could be numeric or character (It will use to print in the PDF)

# File is a list of reports from C-Lock inc.
# it could be one or multiples depend on the number of units
file <- list(system.file("extdata", "StudyName_FinalReport.xlsx", package = "greenfeedr"))

# By default `finalrep()` plot only methane (CH4), but here we defined to plot "All" gases

# Is it possible to include a file with Farm IDs to use in the reports
# The file structure should be: FarmName | RFID

finalrep(Exp = Experiment,
         Unit = 1,
         Start_Date = "2024-05-13",
         End_Date = "2024-05-25",
         Final_report = file,
         Plot_opt = "All",
         RFID_file = NA)

}
