% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_gfdata.R
\name{get_gfdata}
\alias{get_gfdata}
\title{Download Daily GreenFeed Data via API}
\usage{
get_gfdata(
  user,
  pass,
  exp = NA,
  unit,
  start_date,
  end_date = Sys.Date(),
  save_dir = getwd()
)
}
\arguments{
\item{user}{User name to log in to the GreenFeed system}

\item{pass}{Password to log in to the GreenFeed system}

\item{exp}{Study name or other study identifier. It is used as the file name to save the data}

\item{unit}{GreenFeed unit number(s). If multiple units, they could be in a vector, list, or character as "1,2"}

\item{start_date}{Start date of the study}

\item{end_date}{End date of the study. By default the current date is used}

\item{save_dir}{Directory to save the output file. By default the current working directory is used}
}
\value{
This function does not return any value. It saves the downloaded data as a CSV file in the specified directory.
}
\description{
Downloads from C-Lock Inc. server daily GreenFeed data
    via an application programming interface (API).
    Retrieves data based on specified parameters such as login information,
    date range and units, providing an Excel file with GreenFeed data.
}
\examples{
\dontshow{if (has_credentials()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
# Please replace "your_username" and "your_password" with your actual GreenFeed credentials.
user <- Sys.getenv("API_USER")
pass <- Sys.getenv("API_PASS")
exp <- "StudyName"
start_date <- "2024-01-01"
end_date <- Sys.Date()
save_dir <- tempdir()

# Example with multiple units as a comma-separated string
unit <- "304,305"
get_gfdata(user, pass, exp, unit, start_date, end_date, save_dir)

# Example with a single unit as a numeric value
unit <- 304
get_gfdata(user, pass, exp, unit, start_date, end_date, save_dir)

# Example with units as a vector
unit <- c(304, 305)
get_gfdata(user, pass, exp, unit, start_date, end_date, save_dir)
\dontshow{\}) # examplesIf}
}
