% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{normalize_gfdata}
\alias{normalize_gfdata}
\title{Normalize 'GreenFeed' data with different formats}
\usage{
normalize_gfdata(data)
}
\arguments{
\item{data}{a data frame with preliminary or finalized 'GreenFeed' data}
}
\value{
A standardized data frame
}
\description{
Normalize 'GreenFeed' data in their preliminary and finalized format
}
\examples{
file <- readr::read_csv(system.file("extdata", "StudyName_GFdata.csv", package = "greenfeedr"))
data <- transform_gases(data = file)

}
\keyword{internal}
