% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pellin.R
\name{pellin}
\alias{pellin}
\title{Process Pellet Intakes}
\arguments{
\item{Exp}{Study name}

\item{Unit}{List of the unit number(s) of the GreenFeed}

\item{gcup}{Grams of pellets per cup}

\item{Start_Date}{Start date of the study}

\item{End_Date}{End date of the study}

\item{RFID_file}{The file that contains the RFID of the animals enrolled in the study}
}
\value{
An Excel file with pellet intakes for all animals in the study
}
\description{
`pellin()` processes the feedtimes file, including visits and food drops,
    to calculate pellet intakes per animal for the entire requested period.
    This function aggregates data to provide insights into the feeding behavior
    and pellet consumption of the animals during a study.
}
\examples{
\dontrun{
Exp <- "StudyName"
Unit <- list(1, 2)
# Please replace here with the grams of pellet that fit in one cup (10 drop-test)
gcup <- 34
Start_Date <- "2023-01-01"
End_Date <- "2023-04-01"
RFID_file <- "/Users/RFID_file.csv"

pellin(Exp, Unit, gcup, Start_Date, End_Date, RFID_file)
}

}
