% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_gfdata.R
\name{report_gfdata}
\alias{report_gfdata}
\title{Download and Report GreenFeed Data}
\arguments{
\item{user}{a character string representing the user name to log in to the GreenFeed system}

\item{pass}{a character string representing password to log in to the GreenFeed system}

\item{exp}{a character string representing study name or other study identifier. It is used as the file name to save the data}

\item{unit}{numeric or character vector or list representing one or more GreenFeed unit numbers.}

\item{start_date}{a character string representing the start date of the study (format: "mm/dd/yyyy")}

\item{end_date}{a character string representing the end date of the study (format: "mm/dd/yyyy")}

\item{input_type}{a character string representing type of data ('daily' or 'final' report)}

\item{save_dir}{a character string representing the directory to save the output file.}

\item{plot_opt}{a character string representing the gas(es) to plot ('All', or 'CH4', 'CO2', 'O2', 'H2'. By default CH4 will be used}

\item{rfid_file}{a character string representing the file with individual RFIDs. The order should be col1=FarmName and col2=RFID}

\item{file_path}{list of files with the final report(s) from GreenFeed system}
}
\value{
A CSV file with daily GreenFeed data and a PDF report with a description of the daily or final records.
}
\description{
Generates PDF reports of daily and final GreenFeed data.
    If option "daily" is used, data is retrieved from C-Lock server via an application programming interface (API)
    and generates a PDF report to with number of animals, records, and gas production of the ongoing study.
    But, if option "final" is used, final data should be provided to generates a PDF report
    to evaluate all GreenFeed data obtained from the finalized study.
}
\examples{
\dontshow{if (has_credentials()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
# Please replace "your_username" and "your_password" with your actual GreenFeed credentials.
user <- Sys.getenv("API_USER")
pass <- Sys.getenv("API_PASS")

# The data range must be fewer than 180 days
# Example without rfid_file (by default NA)

report_gfdata(user,
  pass,
  exp = "StudyName",
  unit = 1,
  start_date = "2023-01-01",
  end_date = Sys.Date(),
  input_type = "daily",
  save_dir = tempdir(),
  plot_opt = "All"
)
\dontshow{\}) # examplesIf}
# Create a Final Report in PDF format from the finalized data received from C-Lock Inc.
# Note that Unit could be numeric or character (It will use to print in the PDF)

# File is a list of reports from C-Lock inc.
# it could be one or multiples depend on the number of units
file <- list(system.file("extdata", "StudyName_FinalReport.xlsx", package = "greenfeedr"))

# By default `report_gfdata()` plot only methane (CH4), but here we defined to plot "All" gases

# Is it possible to include a file with Farm IDs to use in the reports
# The file structure should be: FarmName | RFID

report_gfdata(
  user = NA,
  pass = NA,
  exp = "StudyName",
  unit = 1,
  start_date = "2024-05-13",
  end_date = "2024-05-25",
  input_type = "final",
  save_dir = tempdir(),
  plot_opt = "All",
  rfid_file = NA,
  file_path = file
)

}
