% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{transform_gases}
\alias{transform_gases}
\title{Transform gas production}
\usage{
transform_gases(data)
}
\arguments{
\item{data}{a data frame with preliminary or finalized 'GreenFeed' data}
}
\value{
A data frame with the gases transform in L/d
}
\description{
Transform gas production from g/d to L/d
}
\examples{
file <- readr::read_csv(system.file("extdata", "StudyName_GFdata.csv", package = "greenfeedr"))
data <- transform_gases(data = file)

}
\keyword{internal}
