% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viseat.R
\name{viseat}
\alias{viseat}
\title{viseat}
\usage{
viseat(
  Exp = NA,
  Unit = list(NA),
  gcup = 34,
  Start_Date = NA,
  End_Date = NA,
  RFID_file = NA
)
}
\arguments{
\item{Exp}{Study name.}

\item{Unit}{List of the unit number/s of the GreenFeed.}

\item{gcup}{Grams of pellets per cup.}

\item{Start_Date}{Start date of the study.}

\item{End_Date}{End date of the study.}

\item{RFID_file}{The file that contains the RFID of the animals enrolled in the study.}
}
\value{
Table with a summary of drops and visits to the GreenFeed unit/s.
}
\description{
Processing visits and feed times in the GreenFeed.
}
