% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dailyrep.R
\name{dailyrep}
\alias{dailyrep}
\title{dailyrep}
\usage{
dailyrep(
  User = NA,
  Pass = NA,
  Exp = NA,
  Unit = NA,
  Start_Date = NA,
  End_Date = Sys.Date(),
  Dir = getwd(),
  RFID_file = NA
)
}
\arguments{
\item{User}{The user name to log in to GreenFeed system.}

\item{Pass}{The password to log in to GreenFeed system.}

\item{Exp}{Study name.}

\item{Unit}{The unit number/s of the GreenFeed.}

\item{Start_Date}{Start date of the study.}

\item{End_Date}{End date of the study. If not specified, the current date will be used.}

\item{Dir}{Directory to save the output file. If not specified, the current working directory will be used.}

\item{RFID_file}{The file that contains the RFID of the animals in the study.}
}
\value{
An excel file with daily data and a PDF report with description of methane data.
}
\description{
Download, processing, and reporting daily GreenFeed data.
}
