% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dailyrep.R
\name{dailyrep}
\alias{dailyrep}
\title{Download and Report Daily GreenFeed Data}
\arguments{
\item{user}{User name to log in to the GreenFeed system}

\item{pass}{Password to log in to the GreenFeed system}

\item{exp}{Study name}

\item{unit}{The unit number(s) of the GreenFeed. If multiple units, they could be in a vector, list, or character as "1,2"}

\item{start_date}{Start date of the study}

\item{end_date}{End date of the study. By default the current date is used}

\item{save_dir}{Directory to save the output file. By default the current working directory is used}

\item{plot_opt}{Type of gas to plot: All, or CH4, CO2, O2, H2. By default only CH4 will be processed and reported}

\item{RFID_file}{File that contains RFID of the animals in the study}
}
\value{
A CSV file with daily data from GreenFeed unit(s) and
    a PDF report with a description of the daily records
}
\description{
`dailyrep()` downloads daily data using an application programming interface (API)
    and generates a PDF report to verify the correct functionality of GreenFeed unit(s).
    This function retrieves data based on the specified parameters, processes it,
    and outputs a summary report in PDF format. The report includes essential metrics and plots
    to ensure that the GreenFeed units are operating as expected.
}
\examples{
\dontshow{if (has_credentials()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
# Please replace "your_username" and "your_password" with your actual GreenFeed credentials.
user <- Sys.getenv("API_USER")
pass <- Sys.getenv("API_PASS")

# The data range must be fewer than 180 days
# Example without RFID_file (by default NA)

dailyrep(user,
         pass,
         exp = "StudyName",
         unit = 1,
         start_date = "2023-01-01",
         end_date = Sys.Date(),
         save_dir = tempdir(),
         plot_opt = "All"
         )
\dontshow{\}) # examplesIf}
}
