% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dailyrep.R
\name{dailyrep}
\alias{dailyrep}
\title{dailyrep}
\arguments{
\item{User}{The user name to log in to GreenFeed system.}

\item{Pass}{The password to log in to GreenFeed system.}

\item{Exp}{Study name.}

\item{Unit}{The unit number/s of the GreenFeed.}

\item{Start_Date}{Start date of the study.}

\item{End_Date}{End date of the study. If not specified, the current date will be used.}

\item{Dir}{Directory to save the output file. If not specified, the current working directory will be used.}

\item{RFID_file}{The file that contains the RFID of the animals in the study.}
}
\value{
An Excel file with the GreenFeed daily data and a PDF report with a description of the gas data.
}
\description{
Download, processing, and reporting daily GreenFeed data.
}
