% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dailyrep.R
\name{dailyrep}
\alias{dailyrep}
\title{dailyrep}
\usage{
dailyrep(
  User = NA,
  Pass = NA,
  Exp = NA,
  Unit = NA,
  Start_Date = NA,
  End_Date = Sys.Date(),
  Output_dir = NA,
  RFID_file = NA
)
}
\arguments{
\item{User}{The user name to log in to GreenFeed system}

\item{Pass}{The password to log in to GreenFeed system}

\item{Exp}{The study name}

\item{Unit}{The number of the GreenFeed unit/s}

\item{Start_Date}{The start date of the study}

\item{End_Date}{The end date of the study. By default is the current Date.}

\item{Output_dir}{The directory to save the report}

\item{RFID_file}{The file that contains the RFID of the animals in the study}
}
\value{
An excel file with daily data and a PDF report with description of GreenFeed data
}
\description{
Download, processing, and reporting daily GreenFeed data
}
