% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/finalrep.R
\name{finalrep}
\alias{finalrep}
\title{finalrep}
\arguments{
\item{Exp}{Study name.}

\item{Unit}{The unit number/s of the GreenFeed.}

\item{Start_Date}{Start date of the study.}

\item{End_Date}{End date of the study.}

\item{Final_report}{List with final report/s generated by C-Lock.}

\item{Plot_opt}{Type of gas to plot: All, CH4, CO2, O2, or H2.}
}
\value{
An excel file with daily data and a PDF report with description of methane data.
}
\description{
Processing and reporting GreenFeed data - final report
}
\examples{
\dontrun{
Exp <- "Test_study"
Unit <- "577,578"
Start_Date <- "2023-01-01"
End_Date <- "2023-04-01"

# Please replace with the final report/s from C-Lock with the finalized GreenFeed data.
## In case you have multiple reports from the same study you can combine them
Final_report <- list("/Users/CLock_finalreport1.xlsx", "/Users/CLock_finalreport2.xlsx")

finalrep(Exp, Unit, Start_Date, End_Date, Final_report, "CH4")
}

}
