% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/finalrep.R
\name{finalrep}
\alias{finalrep}
\title{Report Final GreenFeed Data}
\arguments{
\item{exp}{Study name}

\item{unit}{The unit number(s) of the GreenFeed. If multiple units, they could be in a vector, list, or character as "1,2"}

\item{start_date}{Start date of the study}

\item{end_date}{End date of the study}

\item{final_report}{List with final report(s) generated by C-Lock}

\item{plot_opt}{Type of gas to plot: All, or CH4, CO2, O2, H2. By default only CH4 will be processed and reported}

\item{RFID_file}{The file that contains the RFID of the animals in the study}
}
\value{
A PDF report with description of the final records.
}
\description{
`finalrep()` generates a final report for GreenFeed data, summarizing key metrics
    and providing a comprehensive overview of the collected data.
    This report includes analysis and visualizations to assist in evaluating
    the performance and outcomes of the GreenFeed units.
}
\examples{
# Create a Final Report in PDF format from the finalized data received from C-Lock Inc.
# Define parameters:
Experiment <- "StudyName"

# Note that Unit could be numeric or character (It will use to print in the PDF)

# File is a list of reports from C-Lock inc.
# it could be one or multiples depend on the number of units
file <- list(system.file("extdata", "StudyName_FinalReport.xlsx", package = "greenfeedr"))

# By default `finalrep()` plot only methane (CH4), but here we defined to plot "All" gases

# Is it possible to include a file with Farm IDs to use in the reports
# The file structure should be: FarmName | RFID

finalrep(
  exp = Experiment,
  unit = 1,
  start_date = "2024-05-13",
  end_date = "2024-05-25",
  save_dir = tempdir(),
  final_report = file,
  plot_opt = "All",
  RFID_file = NA
)

}
