% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/finalrep.R
\name{finalrep}
\alias{finalrep}
\title{finalrep}
\usage{
finalrep(
  Exp = NA,
  Unit = NA,
  Start_Date = NA,
  End_Date = NA,
  Final_report = NA
)
}
\arguments{
\item{Exp}{The study name}

\item{Unit}{The number of the GreenFeed unit/s}

\item{Start_Date}{The start date of the study}

\item{End_Date}{The end date of the study}

\item{Final_report}{The final report generate by C-Lock at the end of the study}
}
\value{
An excel file with daily data and a PDF report with description of GreenFeed data
}
\description{
Processing and reporting GreenFeed data - final report
}
