% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/finalrep.R
\name{finalrep}
\alias{finalrep}
\title{finalrep}
\usage{
finalrep(
  Exp = NA,
  Unit = NA,
  Start_Date = NA,
  End_Date = NA,
  Final_report = NA
)
}
\arguments{
\item{Exp}{Study name.}

\item{Unit}{The unit number/s of the GreenFeed.}

\item{Start_Date}{Start date of the study.}

\item{End_Date}{End date of the study.}

\item{Final_report}{Final report generate by C-Lock.}
}
\value{
An excel file with daily data and a PDF report with description of methane data.
}
\description{
Processing and reporting GreenFeed data - final report
}
