% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_api_data.R
\name{get_api_data}
\alias{get_api_data}
\title{Download daily GreenFeed data using API}
\usage{
get_api_data(
  user,
  pass,
  exp,
  unit,
  start_date,
  end_date = Sys.Date(),
  save_dir = getwd()
)
}
\arguments{
\item{user}{User name to log in to the GreenFeed system}

\item{pass}{Password to log in to the GreenFeed system}

\item{exp}{Study name}

\item{unit}{The unit number(s) of the GreenFeed. If multiple units, they could be in a vector, list, or character as "1,2"}

\item{start_date}{Start date of the study}

\item{end_date}{End date of the study. By default the current date is used}

\item{save_dir}{Directory to save the output file. By default the current working directory is used}
}
\value{
A Excel file with the daily data from GreenFeed unit(s)
}
\description{
`get_api_data()` downloads daily data from GreenFeed unit(s)
    using an application programming interface (API). This function
    retrieves data based on specified parameters such as date range and
    unit identifiers, providing a structured data output for further analysis
    or reporting.
}
\examples{
\dontshow{if (has_credentials()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
# Please replace "your_username" and "your_password" with your actual GreenFeed credentials.
user <- Sys.getenv("API_USER")
pass <- Sys.getenv("API_PASS")
exp <- "StudyName"
start_date <- "2023-01-01"
end_date <- Sys.Date()
save_dir <- tempdir()

# Example with multiple units as a comma-separated string
unit <- "304,305"
get_api_data(user, pass, exp, unit, start_date, end_date, save_dir)

# Example with a single unit as a numeric value
unit <- 304
get_api_data(user, pass, exp, unit, start_date, end_date, save_dir)

# Example with units as a vector
unit <- c(304, 305)
get_api_data(user, pass, exp, unit, start_date, end_date, save_dir)
\dontshow{\}) # examplesIf}
}
