% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_api_data.R
\name{get_api_data}
\alias{get_api_data}
\title{Download daily GreenFeed data using API}
\usage{
get_api_data(
  User,
  Pass,
  Exp,
  Unit,
  Start_Date,
  End_Date = Sys.Date(),
  Dir = getwd()
)
}
\arguments{
\item{User}{User name to log in to GreenFeed}

\item{Pass}{Password to log in to GreenFeed}

\item{Exp}{Study name}

\item{Unit}{The unit number(s) of the GreenFeed. If multiple units, they should be separated by a comma.}

\item{Start_Date}{Start date of the study}

\item{End_Date}{End date of the study. If not specified, the current date will be used}

\item{Dir}{Directory to save the output file. If not specified, the current working directory will be used}
}
\value{
A .csv file with daily data from GreenFeed unit(s)
}
\description{
`get_api_data()` downloads daily data from GreenFeed unit(s)
    using an application programming interface (API). This function
    retrieves data based on specified parameters such as date range and
    unit identifiers, providing a structured data output for further analysis
    or reporting.
}
\examples{
\dontshow{if (has_credentials()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
# Please replace "your_username" and "your_password" with your actual GreenFeed credentials.
User <- Sys.getenv("API_USER")
Pass <- Sys.getenv("API_PASS")
Exp <- "StudyName"
Unit <- "304,305" # Specify multiple units as a comma-separated string
Start_Date <- "2023-01-01"
End_Date <- Sys.Date()
Dir <- getwd()

# Example with multiple units as a comma-separated string
get_api_data(User, Pass, Exp, Unit, Start_Date, End_Date, Dir)

# Example with a single unit as a numeric value
Unit <- 304
get_api_data(User, Pass, Exp, Unit, Start_Date, End_Date, Dir)

# Example with units as a vector
Unit <- c(304, 305)
get_api_data(User, Pass, Exp, Unit, Start_Date, End_Date, Dir)

\dontshow{\}) # examplesIf}
}
