% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_api_data.R
\name{get_api_data}
\alias{get_api_data}
\title{get_api_data}
\usage{
get_api_data(User = NA, Pass = NA, Output_dir = NA)
}
\arguments{
\item{User}{The user name to log in to GreenFeed system}

\item{Pass}{The password to log in to GreenFeed system}

\item{Output_dir}{The directory to save the generated file}
}
\value{
An excel file with the daily data from GreenFeed system
}
\description{
Download daily GreenFeed data
}
