% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_api_data.R
\name{get_api_data}
\alias{get_api_data}
\title{get_api_data}
\usage{
get_api_data(
  User = NA,
  Pass = NA,
  Exp = NA,
  Unit = NA,
  Start_Date = NA,
  End_Date = Sys.Date(),
  Dir = getwd()
)
}
\arguments{
\item{User}{The user name to log in to GreenFeed system.}

\item{Pass}{The password to log in to GreenFeed system.}

\item{Exp}{Study name.}

\item{Unit}{The unit number/s of the GreenFeed.}

\item{Start_Date}{Start date of the study.}

\item{End_Date}{End date of the study. If not specified, the current date will be used.}

\item{Dir}{Directory to save the output file. If not specified, the current working directory will be used.}
}
\value{
An excel file with the daily data from GreenFeed system.
}
\description{
Download daily GreenFeed data.
}
