% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_api_data.R
\name{get_api_data}
\alias{get_api_data}
\title{Download daily GreenFeed data using API}
\usage{
get_api_data(
  User = NA,
  Pass = NA,
  Exp = NA,
  Unit = NA,
  Start_Date = NA,
  End_Date = Sys.Date(),
  Dir = getwd()
)
}
\arguments{
\item{User}{User name to log in to GreenFeed}

\item{Pass}{Password to log in to GreenFeed}

\item{Exp}{Study name}

\item{Unit}{The unit number(s) of the GreenFeed}

\item{Start_Date}{Start date of the study}

\item{End_Date}{End date of the study. If not specified, the current date will be used}

\item{Dir}{Directory to save the output file. If not specified, the current working directory will be used}
}
\value{
A .csv file with daily data from GreenFeed unit(s)
}
\description{
`get_api_data()` downloads daily data from GreenFeed unit(s)
    using an application programming interface (API). This function
    retrieves data based on specified parameters such as date range and
    unit identifiers, providing a structured data output for further analysis
    or reporting.
}
\examples{
\dontrun{
# Please replace "your_username" and "your_password" with your actual GreenFeed credentials.
User <- "your_username"
Pass <- "your_password"
Exp <- "StudyName"
Unit <- 1
Start_Date <- "2023-01-01"
End_Date <- Sys.Date()
Dir <- getwd()

get_api_data(User, Pass, Exp, Unit, Start_Date, End_Date, Dir)
}

}
