% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{has_credentials}
\alias{has_credentials}
\title{Check for API Credentials}
\usage{
has_credentials()
}
\description{
`has_credentials()` is a helper function to check
    if the necessary API credentials are available in the environment
}
