% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dailyrep.R, R/get_api_data.R
\name{has_credentials}
\alias{has_credentials}
\title{Check for API Credentials}
\usage{
has_credentials()

has_credentials()
}
\description{
A function to check if the necessary API credentials are available in the environment.

A function to check if the necessary API credentials are available in the environment.
}
