% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pellin.R
\name{pellin}
\alias{pellin}
\title{pellin}
\usage{
pellin(
  Exp = NA,
  Unit = list(NA),
  gcup = 34,
  Start_Date = NA,
  End_Date = NA,
  RFID_file = NA
)
}
\arguments{
\item{Exp}{The study name}

\item{Unit}{List of the GreenFeed unit/s}

\item{gcup}{Grams of pellets per cup}

\item{Start_Date}{The start date of the study}

\item{End_Date}{The end date of the study}

\item{RFID_file}{The file that contains the RFID of the animals in the study}
}
\description{
Processing feedtimes file with intakes
}
