% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_gfdata.R
\name{process_gfdata}
\alias{process_gfdata}
\title{process_gfdata}
\arguments{
\item{file}{File with GreenFeed data}

\item{Start_Date}{Start date of the study}

\item{End_Date}{End date of the study}

\item{input_type}{Data could be from daily or final report: "daily" or "final"}

\item{param1}{Number of records a day}

\item{param2}{Number of days with records}
}
\value{
Tables with processed data
}
\description{
Process daily data from GreenFeed.
}
