% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_gfdata.R
\name{process_gfdata}
\alias{process_gfdata}
\title{Process Daily and Final GreenFeed Data}
\arguments{
\item{file}{File with GreenFeed data in daily or final format}

\item{Start_Date}{Start date of the study}

\item{End_Date}{End date of the study}

\item{input_type}{Input data could be from daily or final report: "daily" or "final"}

\item{param1}{Number of records a day}

\item{param2}{Number of days with records}
}
\value{
Datasets with GreenFeed processed data in a daily and weekly format
}
\description{
`process_gfdata()` processes and computes daily and weekly averages
    for daily and final GreenFeed data, organizing it into structured datasets.
    This function handles data cleaning, aggregation, and summarization to facilitate
    further analysis and reporting.
}
\examples{
\dontrun{
Start_Date <- "2024-01-22"
End_Date <- "2024-03-08"
input_type <- "final"
file <- system.file("extdata", "StudyName_FinalReport.xlsx", package = "greenfeedr")
data <- process_gfdata(file, Start_Date, End_Date, input_type, param1 = 2, param2 = 3)
}

}
