% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_gfdata.R
\name{process_gfdata}
\alias{process_gfdata}
\title{Process Daily and Final GreenFeed Data}
\arguments{
\item{data}{Data frame with daily or finalized GreenFeed data}

\item{start_date}{Start date of the study}

\item{end_date}{End date of the study}

\item{param1}{Number of records per day to be consider for analysis}

\item{param2}{Number of days with records per week to be consider for analysis}

\item{min_time}{Minimum number of minutes for a records to be consider for analysis. By default min_time is 2}
}
\value{
A list of two data frames:
  \item{daily_data}{Data frame with daily processed GreenFeed data}
  \item{weekly_data}{Data frame with weekly processed GreenFeed data}
}
\description{
Processes and computes daily and weekly averages
    for daily and final GreenFeed data, organizing it into two data sets.
    Handles data filtering, aggregation, and summarization to facilitate
    further analysis and reporting.
}
\examples{
file <- system.file("extdata", "StudyName_GFdata.csv", package = "greenfeedr")
data <- readr::read_csv(file)

gf_data <- process_gfdata(data,
  start_date = "2024-05-13",
  end_date = "2024-05-25",
  param1 = 2,
  param2 = 3
)
head(gf_data)

}
