% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_gfdata.R
\name{process_gfdata}
\alias{process_gfdata}
\title{Process daily and final GreenFeed data}
\arguments{
\item{file}{File with GreenFeed data in daily or final format}

\item{Start_Date}{Start date of the study}

\item{End_Date}{End date of the study}

\item{input_type}{Input data could be from daily or final report: "daily" or "final"}

\item{param1}{Number of records a day}

\item{param2}{Number of days with records}
}
\value{
Datasets with GreenFeed processed data in a daily and weekly format
}
\description{
Process daily and final GreenFeed data into structured datasets.
}
\examples{
\dontrun{
Start_Date <- "2024-01-22"
End_Date <- "2024-03-08"
input_type <- "final"
file <- system.file("extdata", "StudyName_FinalReport.xlsx", package = "greenfeedr")
data <- process_gfdata(file, Start_Date, End_Date, input_type, param1 = 2, param2 = 3)
}

}
