% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{process_rfid_data}
\alias{process_rfid_data}
\title{Process RFID Data}
\usage{
process_rfid_data(rfid_file)
}
\arguments{
\item{rfid_file}{Path or data frame containing RFID data.}
}
\description{
`process_rfid_data()` processes RFID data from animals in the study.
}
