% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_gfdata.R
\name{report_gfdata}
\alias{report_gfdata}
\title{Download and Report GreenFeed Data}
\arguments{
\item{user}{User name to log in to the GreenFeed system}

\item{pass}{Password to log in to the GreenFeed system}

\item{exp}{Study name}

\item{unit}{The unit number(s) of the GreenFeed. If multiple units, they could be in a vector, list, or character as "1,2"}

\item{start_date}{Start date of the study}

\item{end_date}{End date of the study. By default the current date is used}

\item{input_type}{Input data could be from daily or final report: daily or final}

\item{save_dir}{Directory to save the output file. By default the current working directory is used}

\item{plot_opt}{Type of gas to plot: All, or CH4, CO2, O2, H2. By default only CH4 will be processed and reported}

\item{RFID_file}{File that contains RFID of the animals in the study}

\item{file}{List of files with final report from GreenFeed}
}
\value{
An Excel file with daily data from GreenFeed unit(s) and
    a markdown report with a description of the daily or final records
}
\description{
`report_gfdata()` generates markdown reports of daily and final data.
    If option "daily" is used, data is retrieve from C-Lock server using an application programming interface (API)
    and generates a markdown report to check correct functionality of GreenFeed unit(s).
    But, if option "final" is used, final data retrieved from C-Lock web interface should be provided to
    generates a markdown report to evaluate the data obtained from the study.
}
\examples{
\dontshow{if (has_credentials()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
#Please replace "your_username" and "your_password" with your actual GreenFeed credentials.
user <- Sys.getenv("API_USER")
pass <- Sys.getenv("API_PASS")

#The data range must be fewer than 180 days
#Example without RFID_file (by default NA)

report_gfdata(user,
              pass,
              exp = "StudyName",
              unit = 1,
              start_date = "2023-01-01",
              end_date = Sys.Date(),
              input_type = "daily",
              save_dir = tempdir(),
              plot_opt = "All"
              )
\dontshow{\}) # examplesIf}
#Create a Final Report in PDF format from the finalized data received from C-Lock Inc.
#Note that Unit could be numeric or character (It will use to print in the PDF)

#File is a list of reports from C-Lock inc.
#it could be one or multiples depend on the number of units
file <- list(system.file("extdata", "StudyName_FinalReport.xlsx", package = "greenfeedr"))

#By default `report_gfdata()` plot only methane (CH4), but here we defined to plot "All" gases

#Is it possible to include a file with Farm IDs to use in the reports
#The file structure should be: FarmName | RFID

report_gfdata(user = NA,
              pass = NA,
              exp = "StudyName",
              unit = 1,
              start_date = "2024-05-13",
              end_date = "2024-05-25",
              input_type = "final",
              save_dir = tempdir(),
              plot_opt = "All",
              RFID_file = NA,
              file = file
              )

}
