% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viseat.R
\name{viseat}
\alias{viseat}
\title{Check GreenFeed Visits}
\arguments{
\item{file}{File with feedtimes from C-Lock. If multiple files are provided, units should be in the same order}

\item{unit}{List of the unit number(s) of the GreenFeed. If multiple files are provided, units should be in the same order}

\item{start_date}{Start date of the study}

\item{end_date}{End date of the study}

\item{RFID_file}{The file that contains the RFID of the animals enrolled in the study. The order should be col1=FarmName and col2=RFID}
}
\value{
A list with two data farmes, one with visits per day and one with visits per animal
}
\description{
`viseat()` processes GreenFeed visits. Generate description of
     animals not visiting the GreenFeed for the time period requested.
}
\examples{
# You should provide the folder where is 'feedtimes' file.
# it could be a list of files if you have data from multiple units to combine
file <- list(system.file("extdata", "feedtimes.csv", package = "greenfeedr"))

# If the user include an RFID file, the structure should be in col1 the farmname or visualID, and
# col2 the RFID or TAG_ID. The file could be save in different formats (.xlsx, .csv, or .txt).
RFIDs <- system.file("extdata", "RFID_file.txt", package = "greenfeedr")

data <- viseat(file,
  unit = 1,
  start_date = "2024-05-13",
  end_date = "2024-05-25",
  RFID_file = RFIDs
)

}
