% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viseat.R
\name{viseat}
\alias{viseat}
\title{Processing visits and feed times in the GreenFeed}
\arguments{
\item{Exp}{Study name}

\item{Unit}{List of the unit number/s of the GreenFeed}

\item{Start_Date}{Start date of the study}

\item{End_Date}{End date of the study}

\item{RFID_file}{The file that contains the RFID of the animals enrolled in the study}
}
\value{
Table with a summary of drops and visits to the GreenFeed unit/s
}
\description{
Processing visits and feed times in the GreenFeed
}
\examples{
\dontrun{
Exp <- "Test_study"
Unit <- list("577", "578")
Start_Date <- "2023-01-01"
End_Date <- "2023-04-01"
RFID_file <- "/Users/RFID_file.csv"

viseat(Exp, Unit, Start_Date, End_Date, RFID_file)
}

}
